# src/docs/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Lava Rapido API",
    "description": "API do banco de dados",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/servicos/": {
      "get": {
        "description": "Busca todos os serviços",
        "responses": {
          "200": {
            "description": "Retorna lista de serviços",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "servico_id": {
                    "type": "number",
                    "example": 1
                  },
                  "nome": {
                    "type": "string",
                    "example": "nome do servico"
                  },
                  "descricao": {
                    "type": "string",
                    "example": "descricao do servico"
                  },
                  "valor_base": {
                    "type": "number",
                    "example": 20
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "description": "Cria um servico",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "nome do servico"
                },
                "descricao": {
                  "type": "string",
                  "example": "descricao do servico"
                },
                "valor_base": {
                  "type": "number",
                  "example": 20
                }
              },
              "required": [
                "nome",
                "descricao",
                "valor_base"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registro criado com sucesso",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Registro criado com sucesso."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/servicos/{id}": {
      "get": {
        "description": "Busca um serviço",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna um serviço",
            "schema": {
              "type": "object",
              "properties": {
                "servico_id": {
                  "type": "number",
                  "example": 1
                },
                "nome": {
                  "type": "string",
                  "example": "nome do servico"
                },
                "descricao": {
                  "type": "string",
                  "example": "descricao do servico"
                },
                "valor_base": {
                  "type": "number",
                  "example": 20
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "put": {
        "description": "Edita um servico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "nome do servico"
                },
                "descricao": {
                  "type": "string",
                  "example": "descricao do servivo"
                },
                "valor_base": {
                  "type": "number",
                  "example": 20
                },
                "servico_id": {
                  "type": "number",
                  "example": 1
                }
              },
              "required": [
                "nome",
                "descricao",
                "valor_base",
                "servico_id"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registro atualizado com sucesso",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Registro atualizado com sucesso."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "description": "Deleta um servico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "servico deletado",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Registro deletado com sucesso."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
    "/pagamentos/": {
      "get": {
        "description": "Busca todos os pagamentos",
        "responses": {
          "200": {
            "description": "Retorna lista de pagamentos",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "pagamento_id": {
                    "type": "number",
                    "example": 1
                  },
                  "os_id": {
                    "type": "number",
                    "example": 1
                  },
                  "forma_pagamento": {
                    "type": "string",
                    "example": "pix"
                  },
                  "valor_pago": {
                    "type": "string",
                    "example": "49.9"
                  },
                  "data_pagamento": {
                    "type": "string",
                    "example": "25-09-25"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "description": "Cria um valor pago ",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "forma_pagamento": {
                  "type": "string",
                  "example": "pix"
                },
                "valor_pago": {
                  "type": "string",
                  "example": "49.9"
                },
                "data_pagamento": {
                  "type": "string",
                  "example": "25-09-25"
                }
              },
              "required": [
                "forma_pagamento",
                "valor_pago",
                "data_pagamento"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registro criado com sucesso",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Registro criado com sucesso."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/pagamentos/{id}": {
      "get": {
        "description": "Busca um valor pago",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna um valor pago",
            "schema": {
              "type": "object",
              "properties": {
                "pagamento_id": {
                  "type": "number",
                  "example": 1
                },
                "os_id": {
                  "type": "number",
                  "example": 1
                },
                "forma_pagamento": {
                  "type": "string",
                  "example": "pix"
                },
                "valor_pago": {
                  "type": "string",
                  "example": "49.9"
                },
                "data_pagamento": {
                  "type": "string",
                  "example": "25-09-25"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "put": {
        "description": "Edita um pagamento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "forma_pagamento": {
                  "type": "string",
                  "example": "pix"
                },
                "valor_pago": {
                  "type": "string",
                  "example": "49.9"
                },
                "data_pagamento": {
                  "type": "string",
                  "example": "25-09-25"
                }
              },
              "required": [
                "forma_pagamento",
                "valor_pago",
                "data_pagamento"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registro atualizado com sucesso",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Registro atualizado com sucesso."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "description": "Deleta um valor pago",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Registro deletado",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Registro deletado com sucesso."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    }
  }
}